+++ game/core/techs.php
@@ -129,6 +129,8 @@ $resourcemap = array ( GID_RC_METAL, GID_RC_CRYSTAL, GID_RC_DEUTERIUM, GID_RC_EN
 $scoreResources = array(GID_RC_METAL, GID_RC_CRYSTAL, GID_RC_DEUTERIUM);
 // Transported resources (usually energy and DM are not transported)
 $transportableResources = array(GID_RC_METAL, GID_RC_CRYSTAL, GID_RC_DEUTERIUM);
+// Resources that convert into debris fields
+$debrisResources = array (GID_RC_METAL, GID_RC_CRYSTAL);
 $prodPriority = array (GID_RC_ENERGY, GID_RC_METAL, GID_RC_CRYSTAL, GID_RC_DEUTERIUM);
 
 // Level 1 cost.
